--- features/cmake/1_export_targets_no_multiconfig/bye/conanfile.py
@@ -1,25 +1,0 @@
-from conans import ConanFile, CMake
-
-
-class ByeConan(ConanFile):
-    name = "bye"
-    version = "1.0"
-    settings = "os", "compiler", "build_type", "arch"
-    options = {"shared": [True, False]}
-    default_options = "shared=False"
-    generators = "cmake_paths"
-    exports_sources = "src/*"
-    requires = "hello/1.0@user/channel"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.definitions["CMAKE_TOOLCHAIN_FILE"] = "conan_paths.cmake"
-        cmake.configure(source_folder="src")
-        cmake.build()
-
-    def package(self):
-        cmake = CMake(self)
-        cmake.install()
-
-    def package_info(self):
-        self.cpp_info.libs = ["bye"]
