--- features/cmake/1_export_targets_no_multiconfig/bye/src/byeConfig.cmake
@@ -1,1 +1,0 @@
-include("${CMAKE_CURRENT_LIST_DIR}/byeTargets.cmake")
