--- features/cmake/1_export_targets_no_multiconfig/bye/test_package/CMakeLists.txt
@@ -1,8 +1,0 @@
-cmake_minimum_required(VERSION 2.8.12)
-project(PackageTest CXX)
-
-include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
-conan_basic_setup()
-
-add_executable(example example.cpp)
-target_link_libraries(example ${CONAN_LIBS})
