--- features/cmake/1_export_targets_no_multiconfig/bye/test_package/example.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-#include "bye.h"
-
-int main() {
-    bye();
-}
