--- features/cmake/1_export_targets_no_multiconfig/hello/conanfile.py
@@ -1,23 +1,0 @@
-from conans import ConanFile, CMake
-
-
-class HelloConan(ConanFile):
-    name = "hello"
-    version = "1.0"
-    settings = "os", "compiler", "build_type", "arch"
-    options = {"shared": [True, False]}
-    default_options = "shared=False"
-    generators = "cmake"
-    exports_sources = "src/*"
-
-    def build(self):
-        cmake = CMake(self)
-        cmake.configure(source_folder="src")
-        cmake.build()
-
-    def package(self):
-        cmake = CMake(self)
-        cmake.install()
-
-    def package_info(self):
-        self.cpp_info.libs = ["hello"]
