--- features/cmake/2_export_targets_multiconfig/bye/src/bye.cpp
@@ -1,15 +1,0 @@
-#include <iostream>
-#include "bye.h"
-#include "hello.h"
-
-void bye(){
-
-    hello();
-
-    #ifdef NDEBUG
-    std::cout << "bye World Release!" <<std::endl;
-    #else
-    std::cout << "bye World Debug!" <<std::endl;
-    #endif
-
-}
