--- features/cmake/2_export_targets_multiconfig/bye/src/bye.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#ifdef WIN32
-  #define HELLO_EXPORT __declspec(dllexport)
-#else
-  #define HELLO_EXPORT
-#endif
-
-HELLO_EXPORT void bye();
