--- features/cmake/2_export_targets_multiconfig/hello/src/hello.cpp
@@ -1,10 +1,0 @@
-#include <iostream>
-#include "hello.h"
-
-void hello(){
-    #ifdef NDEBUG
-    std::cout << "Hello World Release!" <<std::endl;
-    #else
-    std::cout << "Hello World Debug!" <<std::endl;
-    #endif
-}
