--- features/cmake/2_export_targets_multiconfig/hello/src/helloConfig.cmake
@@ -1,1 +1,0 @@
-include("${CMAKE_CURRENT_LIST_DIR}/helloTargets.cmake")
