+++ features/cmake/find_package/exported_targets_multiconfig/README.md
@@ -1,0 +1,64 @@
+Using unobtrusive CMake exported targets with Conan and find_package
+=====================================================================
+
+In this example
+---------------
+
+We are using CMake to declare, export and install the targets of a couple of libraries, and using Conan to manage
+the packages and using these libraries in a project. Everything is based on the ``find_package`` feature and working with 
+multiconfig packages (Release/Debug) to avoid installing the dependencies every time we change from Debug to Release in
+the Visual Studio IDE.
+
+ - We have a "hello" library consumed by a "bye" library and a project that uses the "bye" function.
+ - The "hello" library exports its own ``helloTargets.cmake`` files, everything is pure CMake, every time we call 
+   CMake with a different build type CMake exports a ``helloTargets-{build_type}.cmake`` script.
+ - The conan recipe uses the same binary package for any ``build_type``, by removing the ``build_type`` setting.
+ - The conan recipe at the ``build`` method invokes CMake both for Release and Debug, generating for the same package.
+ - The same for the "bye" library.
+ 
+
+Some useful details of this example
+------------------------------------
+
+ - The ``CMakeLists.txt`` files for ``bye`` and ``hello`` are quite similar, the same code except the library name and
+ dependencies are changed in the first part. So you could use a similar ``CMakeLists.txt`` if you want to follow this 
+ approach.
+ - The ``helloConfig.cmake`` and ``byeConfig.cmake`` are generated by hand, almost only included the autogenerated 
+ targets:
+ 
+ ```
+ include(CMakeFindDependencyMacro)
+ find_dependency(hello)
+
+ include("${CMAKE_CURRENT_LIST_DIR}/byeTargets.cmake")
+ ```
+
+ - The recipes are not even using ``package_info()``  method for the consumers
+
+Pros
+----
+
+- If you are going to consume your packages only with CMake, this is a totally unobtrusive and decoupled mechanism.
+- You don't need to change anything in the ``CMakeLists.txt`` related to Conan.
+- The multi-config mechanism works good, it is very comfortable, only one ``conan install`` for the dependencies.
+
+ 
+Cons
+----
+ 
+ - You cannot consume Conan packages with other mechanism than ``CMake find_package``, no ``package_info`` method has been
+ declared. The logic for the consumer lives exclusively in the ``.cmake`` files installed by CMake, so if I want to use
+ any package with other build system like ``Autotools`` I won't be able to do it.
+ 
+ - You need to build packages containing all the build_types in the same Conan binary package. The ``find_package``
+ mechanism will find only one ``XXXConfig.cmake`` file, so all the ```XXXTarget-{build_type}``` have to be together.
+ 
+ - You have to take care of adjusting the runtime in the CMakeLists.txt to avoid conflicts.
+ 
+ 
+How to try it
+-------------
+
+ - Open the "run.bat" to see the steps of the example. If you run it, it will create the packages for "hello" and "bye" 
+ and will build the "project" changing the build_type and verifying that the linked dependencies are correct for the selected build_type.
+
