+++ features/cmake/find_package/exported_targets_multiconfig/hello/conanfile.py
@@ -1,15 +1,14 @@
 from conans import ConanFile, CMake
 
 
-class ByeConan(ConanFile):
-    name = "bye"
+class HelloConan(ConanFile):
+    name = "hello"
     version = "1.0"
     settings = "os", "compiler", "arch"
     options = {"shared": [True, False]}
     default_options = "shared=False"
-    generators = "cmake_paths"
+    generators = "cmake"
     exports_sources = "src/*"
-    requires = "hello/1.0@user/channel"
 
     def configure(self):
         # it is also necessary to remove the VS runtime
@@ -19,12 +18,6 @@ class ByeConan(ConanFile):
     def build(self):
         for bt in ("Debug", "Release"):
             cmake = CMake(self, build_type=bt)
-            cmake.definitions["CMAKE_TOOLCHAIN_FILE"] = "conan_paths.cmake"
-            cmake.definitions["CMAKE_FIND_DEBUG_MODE"] = "ON"
-
             cmake.configure(source_folder="src")
             cmake.build()
             cmake.install()
-
-    def package_info(self):
-        self.cpp_info.libs = ["bye"]
